% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcplot.R
\name{pcplot}
\alias{pcplot}
\title{Principal Component Plots}
\usage{
pcplot(data, xpc, ypc, cor = TRUE, control = FALSE, group = NA,
  labelled = FALSE)
}
\arguments{
\item{data}{A set of predictor values}

\item{xpc}{Principal component to plot along the x-axis}

\item{ypc}{Principal compenent to plot along y-axis}

\item{cor}{Indicates to use correlation for PCA. Defaults to TRUE}

\item{control}{Indicates if user wants to customize ggplot rather than the default setting. Defaults to FALSE.}

\item{group}{A vector indicating which group the observation belongs to. Defaults to NA.}

\item{labelled}{Indicates whether or not to label the observations. Defaults to FALSE.}
}
\description{
Takes a data set and plots principal components against each other with ggplot.
}
\examples{

library(ggplot2)
data("USArrests")
pcplot(USArrests, 1, 2, labelled = TRUE)

data("longley")
longley$Group <- ifelse(longley$Year <= 1954, "a", "b")
pcplot(longley[,-c(6:8)], 1, 2, group = longley$Group, labelled = TRUE)
}
\keyword{PCA}
\keyword{princomp}
